Add build-vm subcommand to nh os.
This commit adds the `nh os build-vm` subcommand to build a virtual
machine activation script instead of a full system.

This includes a new option --with-bootloader/-B that applies to just
build-vm, to build a VM with a bootloader.

Conditionally run nvd only if the system hostname and target hostnames
match.

See https://github.com/nix-community/nh/pull/208