clean: better interface for commands; passthru `--max`; optimise
Renames `--nogc` and `--nogcroots` to `--no-gc` and `--no-gcroots`
respectively. This is a change in the name of consistency, and should be
handled gracefully thanks to Clap's aliases. Though, you are encouraged
to update your scripts or aliases in case those are deprecated in the
future.

Additionally, `--optimise` and `--max` flags have been introduced. You
may use the former to optimise your Nix store right after a GC and the
latter to fine-grain `nix store gc`'s behaviour.

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: I6a6a69648ffe8a8abe759f773f83b1031e137981